+++ src/addons/Collisions/responses.ts
